# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task Manager API",
    "version": "1.0.0",
    "description": "API para gestión de tareas con Firebase"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor local"
    }
  ],
  "paths": {
    "/tasks": {
      "get": {
        "summary": "Obtener todas las tareas",
        "responses": {
          "200": {
            "description": "Lista de tareas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor"
          }
        }
      },
      "post": {
        "summary": "Crear nueva tarea",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tarea creada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          }
        }
      }
    },
    "/tasks/{id}": {
      "put": {
        "summary": "Actualizar tarea",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskInput"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Tarea actualizada"
          },
          "404": {
            "description": "Tarea no encontrada"
          }
        }
      },
      "delete": {
        "summary": "Eliminar tarea",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Tarea eliminada"
          },
          "404": {
            "description": "Tarea no encontrada"
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Verificar estado del servicio",
        "responses": {
          "200": {
            "description": "Servicio saludable",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "healthy"
                    },
                    "database": {
                      "type": "string",
                      "example": "connected"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Servicio no saludable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "ABC123def456"
          },
          "title": {
            "type": "string",
            "example": "Comprar leche"
          },
          "description": {
            "type": "string",
            "example": "Leche descremada"
          },
          "completed": {
            "type": "boolean",
            "example": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TaskInput": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": {
            "type": "string",
            "example": "Comprar pan"
          },
          "description": {
            "type": "string",
            "example": "Pan integral"
          },
          "completed": {
            "type": "boolean",
            "example": false
          }
        }
      }
    }
  }
}